scripts/requirements-secrets.txt
pynacl>=1.5.0
requests>=2.31.0

